+++ src/OpenTelemetry.Resources.Azure/CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Unreleased
 
 * Add `cloud.account.id` and `azure.resource_group.name` attributes to the Azure
+  VM resource detector.
+  ([#5195](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5195))
+
+## 1.18.0-beta.2
+
+Released 2026-Sep-10
+
+* Add `cloud.account.id` and `azure.resource_group.name` attributes to the Azure
   App Service resource detector.
   ([#5138](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5138))
 
@@ -21,9 +29,13 @@
   requests to the Azure Instance Metadata Service.
   ([#5167](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5167))
 
-* Add `cloud.account.id` and `azure.resource_group.name` attributes to the Azure
-  VM resource detector.
-  ([#5195](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5195))
+* Add `host.image.id`, `host.image.name` and `host.image.version` attributes to
+  the Azure VM resource detector.
+  ([#5194](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5194))
+
+* Updated semantic conventions to
+  [v1.44.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.44.0/docs/resource/host.md).
+  ([#5194](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5194))
 
 ## 1.18.0-beta.1